requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
langchain
langchain-community
langchain-text-splitters
pypdf
sentence-transformers